requirements.txt
Django==1.6.2
Pillow==2.4.0
South==0.8.4
django-braces==1.4.0
django-compressor==1.3
mysql-python==1.2.5
easy-thumbnails==2.0.1
python-memcached==1.53
requests==2.3.0
